requirements.txt
Levenshtein==0.25.1
PyDriller==1.15.5
GitPython==3.1.43
unidiff==0.7.5
openai==1.27.0
gitdb==4.0.11
typing==3.7.4.3